declare id  manifest as Long

+++ src/main/java/com/douane/repository/ManifestRepository.java
@@ -11,16 +11,15 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.douane.entities.Manifest;
 
-public interface ManifestRepository extends JpaRepository<Manifest, Integer>{
+public interface ManifestRepository extends JpaRepository<Manifest, Long>{
 	
 	@Query("SELECT m FROM Manifest m WHERE m.flag= false ")
 	public List<Manifest> getDataNoMarked() ; 
 	
-	
 	@Modifying
 	@Transactional
 	@Query(value = "update Manifest  set  flag='t' where id between :start and  :end  ", nativeQuery = true)
-	public void setMareked(@Param("start") int start ,@Param("end") int end  ) ; 
+	public void setMareked(@Param("start") Long start ,@Param("end") Long end  ) ; 
 	
 	@Query("Select Count(*) from Manifest WHERE  flag  = false ")
 	public int getCount () ;
